Paint calculator
This file gives the user an option to which paint quality they intend to purchase. Then it asks the user to input the number of coats and the surface area of the walls in meter squared. Then it outputs the total cans needed per litre, the coverage of the paint quality chosen in meter squared and the total cost of the paint job.